docs: add extension naming conventions, Amctl project creation, and refactor pyproject.toml and utils
